_legacy/backend_python_v1/requirements.txt
pydantic>=2.0.0
openai>=1.0.0
anthropic>=0.18.0
mss
pyautogui
fastapi
uvicorn
python-dotenv
Pillow
